Fix/clean code

+++ obligatorio_parque_tematico_virtual/TestApi/TicketControllerTest.cs
@@ -1,17 +1,10 @@
-using System;
-using System.Collections.Generic;
-using System.Data.Common;
-using System.Linq;
 using System.Net;
-using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
 using System.Text.Json;
-using System.Threading.Tasks;
 using Domain;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Models.In;
 using Models.Out;
 using IBusinessLogic;
@@ -34,7 +27,6 @@ namespace TestApi
         [TestInitialize]
         public void TestInitialize()
         {
-            // Create shared in-memory connection
             _connection = new SqliteConnection("DataSource=:memory:");
             _connection.Open();
 
@@ -43,7 +35,6 @@ namespace TestApi
                 {
                     builder.ConfigureServices(services =>
                     {
-                        // Remove SQL Server DbContext
                         ServiceDescriptor descriptor = services.SingleOrDefault(
                             d => d.ServiceType == typeof(DbContextOptions<AppDbContext>));
                         if (descriptor != null)
@@ -51,7 +42,6 @@ namespace TestApi
                             services.Remove(descriptor);
                         }
 
-                        // Remove existing IDateTimeLogic registration
                         var dateTimeDescriptor = services.SingleOrDefault(
                             d => d.ServiceType == typeof(IDateTimeLogic));
                         if (dateTimeDescriptor != null)
@@ -59,7 +49,6 @@ namespace TestApi
                             services.Remove(dateTimeDescriptor);
                         }
 
-                        // Remove existing IDateTimeRepository registration
                         var dateTimeRepoDescriptor = services.SingleOrDefault(
                             d => d.ServiceType == typeof(IDateTimeRepository));
                         if (dateTimeRepoDescriptor != null)
@@ -67,7 +56,6 @@ namespace TestApi
                             services.Remove(dateTimeRepoDescriptor);
                         }
 
-                        // Add SQLite DbContext with shared connection
                         services.AddDbContext<AppDbContext>(options =>
                             options.UseSqlite(_connection));
 
@@ -76,7 +64,6 @@ namespace TestApi
                     });
                 });
 
-            // Initialize database schema
             using (IServiceScope scope = _factory.Services.CreateScope())
             {
                 AppDbContext context = scope.ServiceProvider.GetRequiredService<AppDbContext>();
@@ -143,7 +130,7 @@ namespace TestApi
             {
                 VisitorId = registerResult.Id,
                 VisitDate = DateTime.Now.AddDays(7),
-                TicketType = TicketType.General
+                TicketType = (int)TicketType.General
             };
 
             HttpContent purchaseContent = new StringContent(
@@ -161,14 +148,13 @@ namespace TestApi
 
             Assert.IsNotNull(ticketResult);
             Assert.AreEqual(registerResult.Id, ticketResult.VisitorId);
-            Assert.AreEqual(TicketType.General, ticketResult.Type);
+            Assert.AreEqual((int)TicketType.General, ticketResult.Type);
             Assert.AreNotEqual(Guid.Empty, ticketResult.QRCode);
         }
 
         [TestMethod]
         public async Task TestPurchaseTicket_VisitorNotFound()
         {
-            // Register and login a user for authentication
             await _client.PostAsync("/api/auth/register", new StringContent(
                 JsonSerializer.Serialize(new RegisterVisitorRequest
                 {
@@ -187,7 +173,7 @@ namespace TestApi
             {
                 VisitorId = Guid.NewGuid(),
                 VisitDate = DateTime.Now.AddDays(7),
-                TicketType = TicketType.General
+                TicketType = (int)TicketType.General
             };
 
             HttpContent content = new StringContent(
@@ -231,7 +217,7 @@ namespace TestApi
             {
                 VisitorId = registerResult.Id,
                 VisitDate = DateTime.Now.AddDays(-7),
-                TicketType = TicketType.General
+                TicketType = (int)TicketType.General
             };
 
             HttpContent purchaseContent = new StringContent(
@@ -275,7 +261,7 @@ namespace TestApi
             {
                 VisitorId = registerResult.Id,
                 VisitDate = DateTime.Now.AddDays(7),
-                TicketType = TicketType.General
+                TicketType = (int)TicketType.General
             };
 
             HttpContent purchaseContent = new StringContent(
@@ -354,7 +340,7 @@ namespace TestApi
             {
                 VisitorId = registerResult.Id,
                 VisitDate = DateTime.Now.AddDays(7),
-                TicketType = TicketType.General
+                TicketType = (int)TicketType.General
             };
 
             HttpContent purchaseContent1 = new StringContent(
@@ -369,7 +355,7 @@ namespace TestApi
             {
                 VisitorId = registerResult.Id,
                 VisitDate = DateTime.Now.AddDays(14),
-                TicketType = TicketType.EventSpecial,
+                TicketType = (int)TicketType.EventSpecial,
                 EventId = Guid.NewGuid()
             };
 
@@ -400,7 +386,7 @@ namespace TestApi
             {
                 VisitorId = Guid.NewGuid(),
                 VisitDate = DateTime.Now.AddDays(7),
-                TicketType = TicketType.General
+                TicketType = (int)TicketType.General
             };
 
             HttpContent content = new StringContent(
